Make change in Policy, Role middleware, Gate, and can directives
Now admins cant see or perform librarians duty, only the owner can
admins can delete and update only there post
In role middleware, its made to get multiple roles in middleware

+++ app/Http/Controllers/BookController.php
@@ -25,7 +25,8 @@ class BookController extends Controller
     public function create(Book $book)
     {
         Gate::authorize('create', $book);
-        
+
+
         return view('books.create') ;
     }
 
@@ -41,7 +42,7 @@ class BookController extends Controller
         ]) ;
 
         $attribute['user_id'] = Auth::id() ;
-        
+
             if($request->hasFile('BookCover')){
                 $imagepath = $request->file('BookCover')->store('BookCovers', 'public');
                 $attribute['BookCover'] = Storage::url($imagepath);
@@ -52,12 +53,12 @@ class BookController extends Controller
         Book::create($attribute);
 
         $user = Auth::user();
-        if ($user->is_admin){
+        if ($user->role === 'admin'){
             return redirect()->route('admin');
         } else{
             return redirect()->route('books.index');
         }
-        
+
     }
 
     /**
@@ -65,7 +66,7 @@ class BookController extends Controller
      */
     public function show(Book $book)
     {
-        
+
         return view('books.show', ['book' => $book]);
     }
 
@@ -74,21 +75,23 @@ class BookController extends Controller
      */
     public function edit(Book $book)
     {
+        Gate::authorize('update',  $book);
         return view('books.edit', ['book'=> $book]);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Book $book)
+    public function update(Request $request, Book $book, User $user)
     {
+        Gate::authorize('update',  [$book, $user]);
         $attribute = $request->validate([
             'name' => 'required|min:5',
             'description' => 'required',
             'BookCover' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048'
         ]);
 
-        $attribute['user_id'] = Auth::id() ;
+//        $attribute['user_id'] = Auth::id() ;
 
         if($request->hasFile('BookCover')){
             if ($book->BookCover){
@@ -96,19 +99,20 @@ class BookController extends Controller
             }
             $imagepath = $request->file('BookCover')->store('BookCovers', 'public');
             $attribute['BookCover'] = Storage::url($imagepath);
-        } else {
-            $attribute['BookCover'] = null; 
         }
+//        else {
+//            $attribute['BookCover'] = null;
+//        }
 
         $book->update($attribute);
 
         $user = Auth::user();
 
-        if ($user->is_admin){
+        if ($user->role === 'admin'){
             return redirect()->route('admin', ['book' => $book]);
         } else {
             return redirect()->route('books.show', ['book' => $book]);
-        }   
+        }
     }
 
     /**
@@ -116,19 +120,20 @@ class BookController extends Controller
      */
     public function destroy(Book $book , Request $request)
     {
+        Gate::authorize('delete', $book);
             if ($book->BookCover){
                 Storage::disk('public')->delete(str_replace('/storage/', '', $book->BookCover)) ;
         }
         $book->delete();
 
         $user = Auth::user();
-        
-        if ($user->is_admin){
+
+        if ($user->role === 'admin'){
             return redirect()->route('admin', ['book' => $book]);
         } else {
             return redirect()->route('books.index');
-        } 
+        }
 
-        
+
     }
 }
